+++ core/opengate_core/opengate_lib/GateLastVertexSource.cpp
@@ -19,8 +19,9 @@ void GateLastVertexSource::InitializeUserInfo(py::dict &user_info) {
   fN = 0;
 }
 
-double GateLastVertexSource::PrepareNextTime(double current_simulation_time,
-                                             double NumberOfGeneratedEvents) {
+double
+GateLastVertexSource::PrepareNextTime(double current_simulation_time,
+                                      unsigned long NumberOfGeneratedEvents) {
 
   /*
   // If all N events have been generated, we stop (negative time)
